+++ src/medipro/object/camera/FollowingCameraController.java
@@ -28,7 +28,7 @@ public class FollowingCameraController extends CameraController {
      * @param dt    前フレームからの経過時間
      */
     @Override
-    public void update(GameObjectModel model, float dt) {
+    public void update(GameObjectModel model, double dt) {
         FollowingCameraModel _model = (FollowingCameraModel) model;
         if (_model.target.isPresent()) {
             GameObjectModel _target = _model.target.get();
